requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
click==8.3.1
fastapi==0.121.2
h11==0.16.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
pydantic==2.12.4
pydantic_core==2.41.5
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
weasyprint==66.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
brotli==1.2.0
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
cssselect2==0.8.0
fastapi==0.121.2
fonttools==4.60.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
pillow==12.0.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydyf==0.11.0
pyphen==0.17.2
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.49.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
weasyprint==66.0
webencodings==0.5.1
zopfli==0.4.0
stripe==14.0.0  annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
brotli==1.2.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cssselect2==0.8.0
fastapi==0.121.2
fonttools==4.60.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
pillow==12.0.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydyf==0.11.0
pyphen==0.17.2
python-dotenv==1.2.1
python-multipart==0.0.20
requests==2.32.5
sniffio==1.3.1
starlette==0.49.3
stripe==14.0.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
weasyprint==66.0
webencodings==0.5.1
zopfli==0.4.0
